+++ packages/sit-onyx/src/components/OnyxMobileNavButton/OnyxMobileNavButton.ct.tsx
@@ -1,14 +1,14 @@
 import { test } from "../../playwright/a11y";
 import { executeMatrixScreenshotTest, mockPlaywrightIcon } from "../../playwright/screenshots";
-import OnyxNavMobileButton from "./OnyxNavMobileButton.vue";
+import OnyxMobileNavButton from "./OnyxMobileNavButton.vue";
 
 test.describe("Screenshot tests", () => {
   executeMatrixScreenshotTest({
-    name: "Nav mobile button",
+    name: "Mobile nav button",
     columns: ["default", "open"],
     rows: ["default", "hover", "active", "focus-visible"],
     component: (column) => (
-      <OnyxNavMobileButton label="Label" icon={mockPlaywrightIcon} open={column === "open"} />
+      <OnyxMobileNavButton label="Label" icon={mockPlaywrightIcon} open={column === "open"} />
     ),
     beforeScreenshot: async (component, page, column, row) => {
       if (row === "hover") await component.hover();
