+++ packages/sit-onyx/src/index.ts
@@ -55,8 +55,8 @@ export * from "./components/OnyxNavBar/types";
 export { default as OnyxNavItem } from "./components/OnyxNavItem/OnyxNavItem.vue";
 export * from "./components/OnyxNavItem/types";
 
-export { default as OnyxNavMobileButton } from "./components/OnyxNavMobileButton/OnyxNavMobileButton.vue";
-export * from "./components/OnyxNavMobileButton/types";
+export { default as OnyxMobileNavButton } from "./components/OnyxMobileNavButton/OnyxMobileNavButton.vue";
+export * from "./components/OnyxMobileNavButton/types";
 
 export { default as OnyxPageLayout } from "./components/OnyxPageLayout/OnyxPageLayout.vue";
 export * from "./components/OnyxPageLayout/types";
